fix: separate folders for jsdom/node test
