+++ web/__tests__/jsdom/landing.test.tsx
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
-import AppDemo from "../src/components/landing/AppDemo";
-import Features from "../src/components/landing/Features";
-import Hero from "../src/components/landing/Hero";
+import AppDemo from "../../src/components/landing/AppDemo";
+import Features from "../../src/components/landing/Features";
+import Hero from "../../src/components/landing/Hero";
 
 describe("AppDemo Section", () => {
   it("renders heading, highlights, and app image", () => {
