+++ web/jest.config.ts
@@ -1,20 +1,7 @@
-import nextJest from 'next/jest.js'
+import type { Config } from "jest";
 
-const createJestConfig = nextJest({
-  dir: './',   // path to your Next app
-})
+const config: Config = {
+  projects: ["<rootDir>/jest.jsdom.config.ts", "<rootDir>/jest.node.config.ts"],
+};
 
-/**  Custom overrides for the test env  */
-const customJestConfig = {
-  testEnvironment: 'jsdom',
-  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-  moduleNameMapper: {
-    '^@/(.*)$': '<rootDir>/src/$1',
-  },
-  transformIgnorePatterns: [
-    'node_modules/(?!(next|react|react-dom)/)',
-  ],
-}
-
-/**  Export an async config that Next transforms for Jest  */
-export default createJestConfig(customJestConfig)
+export default config;
